requirements.txt
# GRABIT Backend Dependencies

# Core framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# YouTube downloading
pytube>=15.0.0
yt-dlp>=2023.12.30

# Configuration and environment
python-dotenv>=1.0.0

# WebSocket support
websockets>=12.0

# HTTP client for requests
httpx>=0.25.2
requests>=2.31.0

# Async support
asyncio-throttle>=1.0.2

# File handling
pathlib

# Logging
structlog>=23.2.0

# JSON handling
orjson>=3.9.10

# Development dependencies (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Production server (alternative to uvicorn)
gunicorn>=21.2.0

# Process management
psutil>=5.9.6

# Date/time utilities
python-dateutil>=2.8.2

# URL parsing
urllib3>=2.1.0

# System utilities
shutil

# Math and data processing
numpy>=1.24.4

# Optional: Database support
# sqlalchemy>=2.0.23
# alembic>=1.13.1

# Optional: Redis for caching
# redis>=5.0.1
# aioredis>=2.0.1

# Optional: Monitoring
# prometheus-client>=0.19.0
# sentry-sdk>=1.38.0